move test to test_healthsystem.py
